Shuffle definitions among some bot headers
Working to make navgen code include only from shared/, not from sgame/.
