+++ src/sgame/botlib/nav.h
@@ -35,9 +35,12 @@ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 #ifndef BOTLIB_NAV_H_
 #define BOTLIB_NAV_H_
 
-#include "bot_types.h"
+#include "DetourCommon.h"
+#include "DetourNavMesh.h"
+#include "DetourNavMeshBuilder.h"
+#include "DetourTileCache.h"
+#include "DetourTileCacheBuilder.h"
 #include "fastlz/fastlz.h"
-#include "bot_convert.h"
 
 // should be the same as in rest of engine
 #define STEPSIZE 18
@@ -45,6 +48,32 @@ ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 
 static const int NAVMESHSET_MAGIC = 'M'<<24 | 'S'<<16 | 'E'<<8 | 'T'; //'MSET';
 static const int NAVMESHSET_VERSION = 3;
+
+enum navPolyFlags
+{
+	POLYFLAGS_DISABLED = 0,
+	POLYFLAGS_WALK     = 1 << 0,
+	POLYFLAGS_JUMP     = 1 << 1,
+	POLYFLAGS_POUNCE   = 1 << 2,
+	POLYFLAGS_WALLWALK = 1 << 3,
+	POLYFLAGS_LADDER   = 1 << 4,
+	POLYFLAGS_DROPDOWN = 1 << 5,
+	POLYFLAGS_DOOR     = 1 << 6,
+	POLYFLAGS_TELEPORT = 1 << 7,
+	POLYFLAGS_CROUCH   = 1 << 8,
+	POLYFLAGS_SWIM     = 1 << 9,
+	POLYFLAGS_ALL      = 0xffff, // All abilities.
+};
+
+enum navPolyAreas
+{
+	POLYAREA_GROUND     = 1 << 0,
+	POLYAREA_LADDER     = 1 << 1,
+	POLYAREA_WATER      = 1 << 2,
+	POLYAREA_DOOR       = 1 << 3,
+	POLYAREA_JUMPPAD    = 1 << 4,
+	POLYAREA_TELEPORTER = 1 << 5,
+};
 
 struct NavMeshSetHeader
 {
@@ -85,93 +114,6 @@ static inline void SwapNavMeshTileHeader( NavMeshTileHeader &header )
 		dtSwapEndian( &header.tileRef );
 	}
 }
-
-static const int NAVMESHCON_VERSION = 2;
-struct OffMeshConnectionHeader
-{
-	int version;
-	int numConnections;
-};
-
-struct OffMeshConnection
-{
-	rVec  start;
-	rVec  end;
-	float radius;
-	unsigned short flag;
-	unsigned char area;
-	unsigned char dir;
-	unsigned int userid;
-};
-
-struct OffMeshConnections
-{
-	static CONSTEXPR int MAX_CON = 128;
-	float  verts[ MAX_CON * 6 ]; // dtOffMeshConnection::pos
-	float  rad[ MAX_CON ]; // dtOffMeshConnection::rad
-	unsigned short flags[ MAX_CON ]; // dtOffMeshConnection::poly
-	unsigned char areas[ MAX_CON ]; // dtOffMeshConnection::flags
-	unsigned char dirs[ MAX_CON ]; // likely dtOffMeshConnection::side
-	unsigned int userids[ MAX_CON ]; // dtOffMeshConnection::userId
-	int      offMeshConCount;
-
-	OffMeshConnections() : offMeshConCount( 0 ) { }
-
-	void reset()
-	{
-		offMeshConCount = 0;
-	}
-
-	void delConnection( int index )
-	{
-		if ( index < 0 || index >= offMeshConCount )
-		{
-			return;
-		}
-
-		for ( int i = index * 6; i < offMeshConCount * 6 - 1; i++ )
-		{
-			verts[ i ] = verts[ i + 1 ];
-		}
-
-		for ( int i = index; i < offMeshConCount - 1; i++ )
-		{
-			rad[ i ] = rad[ i + 1 ];
-			flags[ i ] = flags[ i + 1 ];
-			areas[ i ] = areas[ i + 1 ];
-			dirs[ i ] = dirs[ i + 1 ];
-			userids[ i ] = userids[ i + 1 ];
-		}
-		offMeshConCount--;
-	}
-
-	bool addConnection( const OffMeshConnection &c )
-	{
-		if ( offMeshConCount == MAX_CON )
-		{
-			return false;
-		}
-
-		float *start = &verts[ offMeshConCount * 6 ];
-		float *end = start + 3;
-
-		start[ 0 ] = c.start[ 0 ];
-		start[ 1 ] = c.start[ 1 ];
-		start[ 2 ] = c.start[ 2 ];
-		end[ 0 ] = c.end[ 0 ];
-		end[ 1 ] = c.end[ 1 ];
-		end[ 2 ] = c.end[ 2 ];
-
-		rad[ offMeshConCount ] = c.radius;
-		flags[ offMeshConCount ] = c.flag;
-		areas[ offMeshConCount ] = c.area;
-		dirs[ offMeshConCount ] = c.dir;
-		userids[ offMeshConCount ] = c.userid;
-
-		offMeshConCount++;
-		return true;
-	}
-};
 
 struct FastLZCompressor : public dtTileCacheCompressor
 {
